Conflicts:
	cms/templates/cms/toolbar/items/login.html

+++ menus/menu_pool.py
@@ -155,15 +155,14 @@ class MenuPool(object):
         else:
             cache_keys = CacheKey.objects.get_keys(site_id, language)
         to_be_deleted = cache_keys.distinct().values_list('key', flat=True)
-        cache.delete_many(to_be_deleted)
-        cache_keys.delete()
+        if to_be_deleted:
+            cache.delete_many(to_be_deleted)
+            cache_keys.delete()
 
     def register_menu(self, menu_cls):
-        import os
-        import inspect
         import warnings
-        source_file = os.path.basename(inspect.stack()[1][1])
-        if source_file == 'menu.py':
+
+        if menu_cls.__module__.split('.')[-1] == 'menu':
             warnings.warn('menu.py filename is deprecated, '
                           'and it will be removed in version 3.4; '
                           'please rename it to cms_menus.py', DeprecationWarning)
